requirements.txt
flask==3.0.0
flask-cors==4.0.1
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
facenet-pytorch==2.5.3
torch>=2.0.0
torchaudio>=2.0.0
openai-whisper==20240930
pyannote.audio==3.1.1
soundfile==0.12.1
sqlalchemy==2.0.36
python-dotenv==1.0.1
groq>=1.0.0
